requirements.txt
torch
torchvision
numpy
opencv-python
pillow
matplotlib
tqdm